requirements.txt
faker
psycopg2-binary
python-dotenv
streamlit